Add some raindancing to try and allow turning the "Can't figure manpath" error into a warning during configuration, while still making it a fatal error before any installation happens to avoid leaving people with partial installs.
The non-stupid way to do it would be to make a preinstall-nomanpath
target or some such, and add a dependancy from install to it.  But
cmake doesn't let us do that; good job, guys.  add_custom_command() in
various guises similarly doesn't seem to provide anything useful.  So
we're down to doing crazy magic in an install(CODE) target.  This will
also not work reliably down the road if and when we split up into
multiple directories, because the order of install()'s is explicitly
unspecified in that case; I _think_ it's guaranteed for the single dir
case though.

This is really a mess.