Add new defines in new locations for various installs.  This installs them now in much different places than before.
Major behavioral change from ctwm's perspective is that the
system.ctwmrc is now installed in an examples dir, while at runtime it
looks in an etc dir, which will be not have the system.ctwmrc by
default.  That's OK, because it's just the one that's compiled into
ctwm anyway, so nobody could have a use for it as-is without editing
anyway.