Rearrange this Makefile a little so there's more separation between "different ways of generating X" and "way we default to X".  And switch the defaults to asciidoctor.
This is slightly premature for the manpage, since there's not yet a
released version that does build manpages.  I define this as a
non-issue, since this Makefile isn't used by any normal build
processes, but only possibly matters to devs working on the manual,
and to the release engineer.  The latter and most of the former are
me, and I've got the dev version in my $PATH, and asciidoctor is a
factor of >10 faster than asciidoc.