Just switch to conditionalizing the - and un-expanding TitleButtonShadowDepth for the L hilite window.
This works fine on UseThreeDTitles.  It looks bad on the 2d case, but
the lolite window already did, and it looks worse having them not
match (though you don't notice unless you have UseSunkTitlePixmap set,
otherwise you don't even get the lolite windows).  But this is a step
along the way; more reorganizing is needed to get it right in both
cases.