Add a trivial wrapper to cover an overwhelmingly common case.
LookInList takes a string and a struct containing two other strings,
looks for something in the list matching them, and returns it.
However, in a great many cases, the string and struct are the same
bits from a TwmWindow structure, and in a huge number of those cases,
we don't even care _what_ in the list matched them, just whether
_something_ did.  So make a wrapper that just takes the TwmWindow and
returns bool.