Add options to tell flex to disable bothering with yywrap and yyunput. Include comments with the version numbers this will blow up before; I choose to completely ignore the issue until somebody shows up with a >22 year old version of flex.
The latter will probably clean up some build warnings on some
platforms.  The former may do the same, and additionally sidesteps a
bug in flex 2.6.3 that lead to the macro being defined and causing
troubles when it tried to link since it couldn't find yywrap() (our
faked version not being added due to the #ifdef).  x-ref
<https://github.com/westes/flex/issues/162>