Rearrange some of the conditionals around HTML manual building, so that if we have a pregen'd manual, we use it even if we find asciidoc but not asciidoctor.
Previously, the process was "if(asciidoctor) use it; elseif(asciidoc)
add target but don't use by default; elseif(have pregen) use it;".
This changes it to use the pregen'd file in the elseif(asciidoc) case;
this makes it a little harder to manually call asciidoc during the
main process, but will still install the pregen'd file by default.
That's a fair tradeoff.