const-ify the callback func for RAreaListForeach().
This isn't entirely clearcut; we can easily imagine ways to use it
where we might _want_ to mutate the RArea's as they pass through.
However, no current code does.  So, let's err on the tighter side.  If
we ever grow a non-const user, we can re-evaluate then.

(not quite a complete conversion, as r_layout.c changes for this are
mixed with other stuff there)