Reworking the building a little so main() becomes ctwm_main(), and we link everything into a little static lib, then have a separate main() that just calls that and we link out into the final binary.
This has no functional effect on the final result, but leaves us an
intermediate library we can link in to get access to all the ctwm
funcs in other binaries.  This is of some use in e.g. writing test
programs.

It would be preferable to use cmake object libraries instead of an an
actual static lib, but that requires a newer cmake than we currently
require.  Todo: revisit the question of what we require there and
probably bump up.